requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
autopep8
tzdata
toml
pycodestyle
django-import-export
django-crispy-forms
django-mathfilters
django-ckeditor
django-ckeditor-5
django-taggit
requests
